refactor: densify leftover AI observer/support hosts and ratchet 300-line gates (#4602)
Topic-split remaining 301–399 observer and support modules so Slice E can
drop the 400-line suite ceilings to 300 without dropping diagnostic pins.
